Add support to fetch user details via LDAP
Since the group CR will just accept the list of unique userIDs, these
userIDs will have a corresponding mapping in the respective company's
LDAP which will provide us the following:
- username
- first name
- last name
- email
- display name

The same LDAP can also be leveraged to check the existence of user.

This commit adds support to fetch user details from LDAP and sync it
against the respective backends.

Signed-off-by: vinamra28 <vinjain@redhat.com>