Commit mock method stubs instead of ignoring them
Earlier we planned to gitignore the mock method stubs created by mockgen
to avoid commiting the code which can be autogenerated. This raises an
issue with dependabot as before updating any new deps, it needs to first
build the application and then only it can raise the PR.

Removing `**/mocks/**` from the gitignore.

Signed-off-by: vinamra28 <vinjain@redhat.com>