Abstract Cache operation to a unified store pkg
In order to make operations on user, team and metadata consistent across
different callers, we extracted this out into it's different package so
that each caller follows a unified pattern.

Also prepended prefix keys for each operation so that it's easy to
identify the record type.

Added UTs and fixed linting issues.

Signed-off-by: vinamra28 <vinjain@redhat.com>