Restructure data in redis to store user details
Since redis until now was storing IDs of users and teams against each
backend

user:
```
Key: user:vinjain@redhat.com
Value: {
  "fivetran_fivetran": "user_abc123",
  "rover_rover": "user_xyz789",
  "rhplatformtest_snowflake": "SNOWFLAKE_USER_ID_456",
  "gitlab_gitlab": "gitlab_user_789"
}
```

team:
```
Key: team:dataverse-platform-usernauttest
Value: {
  "fivetran_fivetran": "team_abc123",
  "rover_rover": "team_xyz789",
  "rhplatformtest_snowflake": "SNOWFLAKE_ROLE_ID_456",
  "gitlab_gitlab": "gitlab_group_789"
}
```

while the above is good for basic use cases, in cases where we want to
fetch complete user details such as to which groups a user is part of or
to see how many users are there in a group, we have to depend on the
kubernetes CRs which is another API call to the kube server.

Added following keys in redis for each fetch:

User group store:

```
Key: `user:groups:<email>`
Value: ["group1", "group2", "group3"]
```

team
```
Key: team:<actual_group_name>
Value: {
    "members": ["user1@example.com", "user2@example.com"],
    "backends": {
      "fivetran_fivetran": {"id": "team_123", "name": "fivetran", "type": "fivetran"},
      "rover_rover": {"id": "team_456", "name": "rover", "type": "rover"}
    }
  }
```

Kept previous team store as is to support preloadCache as backends are
going to return transformed name. During reconciliation, if the ID is
not found in the new store, it should fallback to team store populated
by preloadCache and then if data isn't found, it should create.

Signed-off-by: vinamra28 <vinjain@redhat.com>